Replace quiet straight on instruction by a quiet flag that can be applied to other instructions as well.
This is intended for suppressing later on e.g. half left and half right instructions as well, if appropriate, i.e. there's no change of the way's name.

+++ GpsMidGraph/de/ueller/midlet/gps/routing/ConnectionWithNode.java
@@ -11,7 +11,7 @@ public class ConnectionWithNode {
 	public short wayFromConAt=0;
 	public short wayToConAt=0;
 	public byte wayType=0;
-	public byte wayRouteFlags=0;
+	public short wayRouteFlags=0;
 	public byte wayConEndBearing=0;
 	public byte wayConStartBearing=0;
 	public float wayDistanceToNext=Float.MAX_VALUE;
